requirements.txt
torch==2.10.0
torchvision==0.25.0
transformers==5.3.0
datasets==4.6.1
evaluate==0.4.6
rouge_score==0.1.2
nltk==3.9.3
pillow==12.1.1
tqdm==4.67.3
ipywidgets==8.1.8
python-dotenv==1.2.1
transformers[torch]
tensorboard==2.20.0
accelerate==1.13.0
matplotlib==3.10.8